% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.evonet.treedata}
\alias{as.evonet.treedata}
\title{Treedata object to evonet}
\usage{
\method{as.evonet}{treedata}(trd)
}
\arguments{
\item{trd}{A treedata object that contains a phylo slot in evonet format}
}
\value{
An evonet object
}
\description{
Treedata object to evonet
}
