% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.treedata.evonet}
\alias{as.treedata.evonet}
\title{Evonet object to treedata}
\usage{
\method{as.treedata}{evonet}(evo, ...)
}
\arguments{
\item{evo}{An evonet object.}

\item{...}{Unused, for extensibility.}
}
\value{
A treedata object with evonet object in the phylo slot.
}
\description{
A method for converting evonet objects to treedata objects. Please note:
We are currently changing class information to c("phylo", "evonet")
even though the other way around is \emph{correct}. This is because the S4 def.
of treedata does not allow for evonet objects in the phylo slot. At some
point we will resolve this inconsistency.
}
