% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.evonet}
\alias{fortify.evonet}
\title{Fortify evonet objects}
\usage{
\method{fortify}{evonet}(
  model,
  data,
  layout = "rectangular",
  ladderize = TRUE,
  right = FALSE,
  branch.length = "branch.length",
  mrsd = NULL,
  as.Date = FALSE,
  yscale = "none",
  root.position = 0,
  ...
)
}
\arguments{
\item{model}{evonet object to convert to data frame}

\item{data}{original data}

\item{layout}{Currently only 'rectangular' layout is supported}

\item{ladderize}{if TRUE it reorganizes the backbone tree to get an laddered structure}

\item{right}{specifies whether the smallest clade should be on the right or left}

\item{branch.length}{specifies branch length attribute}

\item{mrsd}{specifies mrsd attribute}

\item{as.Date}{specifies as.Date attribute}

\item{yscale}{specifies y scale attribute}

\item{root.position}{specifies root attribute}

\item{...}{}
}
\value{
fortified data frame
}
\description{
Transforms evonet objects into dataframes that can be passed to ggplot2
}
