% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ret.R
\name{geom_ret}
\alias{geom_ret}
\title{Explicit phylogenetic network}
\usage{
geom_ret(
  retcol = "black",
  arrows = FALSE,
  retlinetype = 2,
  rettype = "snake",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{retcol}{reticulation edge colour}

\item{arrows}{if TRUE reticulation edges end with arrows}

\item{retlinetype}{line type aesthetic for reticulations}

\item{rettype}{Either 'straight' or 'snake'. Determines in which style reticulations are drawn}

\item{na.rm}{logical}

\item{...}{additional parameter}
}
\value{
phylogenetic network layer
}
\description{
Explicit phylogenetic network
}
\author{
Yu Guangchuang
}
