% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-utilities.R
\name{getNodeNum}
\alias{getNodeNum}
\title{layoutEqualAngle}
\usage{
getNodeNum(tree)
}
\arguments{
\item{model}{tree object, e.g. phylo or treedata}

\item{branch.length}{set to 'none' for edge length of 1. Otherwise the phylogenetic tree edge length is used.}
}
\value{
tree as data.frame with equal angle layout.
}
\description{
'Equal-angle layout algorithm for unrooted trees'
}
\references{
"Inferring Phylogenies" by Joseph Felsenstein.
}
