% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggarg-package.R
\docType{package}
\name{ggarg-package}
\alias{ggarg}
\alias{ggarg-package}
\title{ggarg: Visualisation of Tree-based Phylogenetic Networks}
\description{
Visualisation and annotation of Ancestral Recombination Graphs and other tree-based phylogenetic networks. It extends the functionality of 'ggtree' and 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grdspcht/ggarg}
  \item Report bugs at \url{https://github.com/grdspcht/ggarg/issues}
}

}
\author{
\strong{Maintainer}: Gerd Specht \email{specht@gea.mpg.de}

}
\keyword{internal}
