% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arg.R
\name{ggarg}
\alias{ggarg}
\title{Adds tree-based network layer}
\usage{
ggarg(arg, retcol = "black", lty = "dashed", rty = "rectangular")
}
\arguments{
\item{arg}{A tree-based network}
}
\value{
ARG plot
}
\description{
Adds tree-based network layer
}
