% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggarg-package.R
\docType{package}
\name{ggret-package}
\alias{ggret}
\alias{ggret-package}
\title{ggret: Visualisation of Tree-based Phylogenetic Networks}
\description{
Visualisation and annotation of Ancestral Recombination Graphs and other tree-based phylogenetic networks. It extends the functionality of 'ggtree' and 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grdspcht/ggarg}
  \item Report bugs at \url{https://github.com/grdspcht/ggarg/issues}
}

}
\author{
\strong{Maintainer}: Gerd Specht \email{gerdspecht1995@gmail.com}

}
\keyword{internal}
