% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggret.R
\name{ggret}
\alias{ggret}
\title{Plot reticulation networks}
\usage{
ggret(
  data,
  mapping = NULL,
  retcol = "black",
  arrows = FALSE,
  retlinetype = 2,
  rettype = "snake",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{data}{either treedata or evonet objects}

\item{mapping}{aesthetic mappings}

\item{retcol}{reticulation edge colour}

\item{arrows}{if TRUE reticulation edges end with arrows}

\item{retlinetype}{line type aesthetic for reticulations}

\item{rettype}{Either 'straight' or 'snake'. Determines in which style reticulations are drawn}

\item{na.rm}{supresses NA value warnings}

\item{...}{additional parameter}
}
\value{
reticulation network plot
}
\description{
Plot reticulation networks
}
