% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClade.R
\name{groupClade}
\alias{groupClade}
\title{group clades}
\usage{
groupClade(data, nodes, cladename, tiponly = FALSE)
}
\arguments{
\item{data}{evonet or treedata object}

\item{nodes}{nodes that define the clade}

\item{cladename}{Name of the clade}

\item{tiponly}{Should only tips be considered?}
}
\description{
group clades
}
