% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arg.R
\name{has_name}
\alias{has_name}
\title{Give a deprecation error, warning, or message, depending on version number.}
\usage{
has_name(x)
}
\arguments{
\item{version}{The last version of ggplot2 where this function was good
(in other words, the last version where it was not deprecated).}

\item{msg}{The message to print.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
