% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_clade.R
\name{modify_ret}
\alias{modify_ret}
\title{Change reticulation coloring}
\usage{
modify_ret(plot, assignto = "donor")
}
\arguments{
\item{plot}{ggplot object}

\item{assignto}{Reticulation color can be assigned to "donor" or "receiver"}
}
\value{
Modified ggplot object
}
\description{
Change reticulation coloring
}
