% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorClade.R
\name{modifyRet}
\alias{modifyRet}
\title{Change reticulation coloring}
\usage{
modifyRet(plot, assignto = "donor")
}
\arguments{
\item{plot}{ggplot object}

\item{assignto}{Reticulation colour can be assigned to "donor" or "receiver"}
}
\value{
Modified ggplot object
}
\description{
Change reticulation coloring
}
