% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbeast.R
\name{read.beast}
\alias{read.beast}
\title{Read Beast2 files with phylogenetic network data}
\usage{
read.beast(file)
}
\arguments{
\item{file}{Beast2 file}
}
\value{
treedata object
}
\description{
Read Beast2 files with phylogenetic network data
}
