% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readbeast.R
\name{read.beast.retnet}
\alias{read.beast.retnet}
\title{read beast files that contain phylogenetic networks}
\usage{
read.beast.retnet(file)
}
\arguments{
\item{file}{Beast2 file with extended newick block}
}
\value{
treedata object
}
\description{
Read Beast2 files with phylogenetic network data
}
