% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2.R
\name{read.enewick}
\alias{read.enewick}
\title{Read Network in Extended Newick Format}
\usage{
read.enewick(file = "", text = NULL)
}
\arguments{
\item{file}{extended newick file}

\item{text}{character string}
}
\value{
evonet object
}
\description{
Read Network in Extended Newick Format
}
