% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2.R
\name{read.enewick2}
\alias{read.enewick2}
\title{Read an extended Newick file (via ape)}
\usage{
read.enewick2(file = "")
}
\arguments{
\item{file}{Path to an extended Newick file.}
}
\value{
ape evonet object
}
\description{
Read an extended Newick file (via ape)
}
\examples{
print("EXAMPLE HERE")
}
