% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_beast.R
\name{read_beast_retnet}
\alias{read_beast_retnet}
\title{read beast files that contain phylogenetic networks}
\usage{
read_beast_retnet(file)
}
\arguments{
\item{file}{Beast2 file with extended newick block}
}
\value{
treedata object
}
\description{
Read Beast2 files with phylogenetic network data
}
