% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_enewick.R
\name{read_enewick}
\alias{read_enewick}
\title{Read Network in Extended Newick Format}
\usage{
read_enewick(file = "", text = NULL)
}
\arguments{
\item{file}{extended newick file}

\item{text}{character string}
}
\value{
evonet object
}
\description{
Read Network in Extended Newick Format
}
