% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retnet}
\alias{retnet}
\title{An example reticulation network}
\format{
\subsection{\code{retnet}}{

An evonet object with 2 reticulations, 4 tips and 8 internal nodes:
\describe{
\item{edge}{Edge matrix that describes the base tree toplogy}
\item{edge.length}{Edge lengths}
\item{Nnode}{Number of internal nodes}
\item{node.label}{Labels for internal nodes. Root node: "r"}
\item{tip.label}{Labels for network tips. (A-D)}
\item{root.edge}{edge identifier of root}
\item{reticulation}{Edge matrix for reticulations}
\item{ret.length}{Reticulation Lengths}
...
}
}
}
\usage{
retnet
}
\description{
A small reticulation network for testing purposes
This object was designed and exported in PhyloSketch.
Edge length values were added by hand.
}
\keyword{datasets}
