% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{retnet_evonet}
\alias{retnet_evonet}
\title{An example reticulation network in evonet format}
\format{
\subsection{\code{retnet_evonet}}{

Evolutionary network with 4 reticulations, the base tree has 20 tips and 27 internal nodes:
\describe{
\item{edge}{Edge matrix that describes the base tree toplogy}
\item{edge.length}{Edge lengths}
\item{Nnode}{Number of internal nodes}
\item{node.label}{Labels for internal nodes. Root node: "r"}
\item{tip.label}{Labels for network tips. (A-D)}
\item{root.edge}{edge identifier of root}
\item{reticulation}{Edge matrix for reticulations}
\item{ret.length}{Reticulation Lengths}
...
}
}
}
\usage{
retnet_evonet
}
\description{
A small reticulation network for testing purposes.
}
\keyword{datasets}
