% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.evonet.treedata}
\alias{as.evonet.treedata}
\title{Treedata object to evonet}
\usage{
\method{as.evonet}{treedata}(x, ...)
}
\arguments{
\item{x}{A treedata object that contains a phylo slot in evonet format}

\item{...}{Unused, for extensibility.}
}
\value{
An evonet object
}
\description{
Treedata object to evonet
}
