% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_arg.R
\name{geom_arg}
\alias{geom_arg}
\title{geom_tree2}
\usage{
geom_arg(layout = "rectangular", retcol = retcol, arrows = FALSE, ...)
}
\arguments{
\item{layout}{one of 'rectangular', 'slanted', 'circular', 'radial' or 'unrooted'}

\item{...}{additional parameter}
}
\value{
tree layer
}
\description{
add tree layer
}
\author{
Yu Guangchuang
}
