% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggret-package.R
\docType{package}
\name{ggret-package}
\alias{ggret-package}
\alias{_PACKAGE}
\alias{ggret-pkg}
\title{ggret: Visualisation of Tree-based Phylogenetic Networks}
\description{
Visualisation and annotation of Ancestral Recombination Graphs and other tree-based phylogenetic networks. It extends the functionality of 'ggtree' and 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grdspcht/ggret}
  \item Report bugs at \url{https://github.com/grdspcht/ggret/issues}
}

}
\author{
\strong{Maintainer}: Gerd Specht \email{gerdspecht1995@gmail.com}

Authors:
\itemize{
  \item Arthur Kocher \email{arthur_kocher@eva.mpg.de} (\href{https://orcid.org/0000-0002-9499-6472}{ORCID}) [research team head]
}

}
\keyword{internal}
