% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggret.R
\name{ggret}
\alias{ggret}
\title{Plot reticulation networks}
\usage{
ggret(
  data,
  mapping = NULL,
  retcol = "black",
  arrows = FALSE,
  retlinetype = 2,
  rettype = "snake",
  na.rm = TRUE,
  mrsd = NULL,
  as.Date = FALSE,
  yscale = "none",
  yscale_mapping = NULL,
  ladderize = TRUE,
  right = FALSE,
  branch.length = "branch.length",
  root.position = 0,
  xlim = NULL,
  layout.params = list(as.graph = TRUE),
  hang = 0.1,
  ...
)
}
\arguments{
\item{data}{either treedata or evonet objects}

\item{mapping}{aesthetic mappings}

\item{retcol}{reticulation edge colour}

\item{arrows}{if TRUE reticulation edges end with arrows}

\item{retlinetype}{line type aesthetic for reticulations}

\item{rettype}{Either 'straight' or 'snake'. Determines in which style reticulations are drawn}

\item{na.rm}{supresses NA value warnings}

\item{...}{additional parameter}
}
\value{
reticulation network plot
}
\description{
Plot reticulation networks
}
