% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupClade.R
\name{groupClade}
\alias{groupClade}
\title{Group clades}
\usage{
groupClade(
  data,
  nodes,
  cladename,
  tiponly = FALSE,
  addtotreedata = FALSE,
  undefinedclades = NA
)
}
\arguments{
\item{data}{Evonet or treedata object.}

\item{nodes}{Nodes that define the clade.}

\item{cladename}{Name of the clade.}

\item{tiponly}{Should only tips be considered?}

\item{addtotreedata}{If TRUE, adds group/clade information to the data table of a treedata object.}

\item{undefinedclades}{Can be changed to rename undefined clades to something more fitting}
}
\description{
Group clades
}
