% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.enewick}
\alias{read.enewick}
\title{Read an extended Newick file}
\usage{
read.enewick(file = "")
}
\arguments{
\item{file}{Path to an extended Newick file.}
}
\value{
ape evonet object
}
\description{
Read an extended Newick file
}
\examples{
print("EXAMPLE HERE")
}
